ACMS-651: Acquia connector form refactored.
