+++ modules/acquia_cms_tour/src/Form/AcquiaConnectorForm.php
@@ -2,26 +2,13 @@
 
 namespace Drupal\acquia_cms_tour\Form;
 
-use Drupal\Core\Extension\InfoParserInterface;
-use Drupal\Core\Extension\ModuleHandlerInterface;
-use Drupal\Core\Form\ConfigFormBase;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\State\StateInterface;
 use Drupal\Core\Url;
-use Drupal\Core\Utility\LinkGeneratorInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Provides a form to configure Acquia Connector.
  */
-final class AcquiaConnectorForm extends ConfigFormBase implements AcquiaDashboardInterface {
-
-  /**
-   * The state service.
-   *
-   * @var \Drupal\Core\State\StateInterface
-   */
-  protected $state;
+final class AcquiaConnectorForm extends AcquiaCMSDashboardBase {
 
   /**
    * Provides module name.
@@ -29,58 +16,6 @@ final class AcquiaConnectorForm extends ConfigFormBase implements AcquiaDashboar
    * @var string
    */
   protected $module = 'acquia_connector';
-
-  /**
-   * The module handler.
-   *
-   * @var \Drupal\Core\Extension\ModuleHandlerInterface
-   */
-  protected $moduleHandler;
-
-  /**
-   * The link generator.
-   *
-   * @var \Drupal\Core\Utility\LinkGeneratorInterface
-   */
-  protected $linkGenerator;
-
-  /**
-   * The info file parser.
-   *
-   * @var \Drupal\Core\Extension\InfoParserInterface
-   */
-  protected $infoParser;
-
-  /**
-   * Constructs a new AcquiaConnectorForm.
-   *
-   * @param \Drupal\Core\State\StateInterface $state
-   *   The state service.
-   * @param \Drupal\Core\Extension\ModuleHandlerInterface $module_handler
-   *   The module handler service.
-   * @param \Drupal\Core\Utility\LinkGeneratorInterface $link_generator
-   *   The link generator.
-   * @param \Drupal\Core\Extension\InfoParserInterface $info_parser
-   *   The info file parser.
-   */
-  public function __construct(StateInterface $state, ModuleHandlerInterface $module_handler, LinkGeneratorInterface $link_generator, InfoParserInterface $info_parser) {
-    $this->state = $state;
-    $this->module_handler = $module_handler;
-    $this->linkGenerator = $link_generator;
-    $this->infoParser = $info_parser;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public static function create(ContainerInterface $container) {
-    return new static(
-      $container->get('state'),
-      $container->get('module_handler'),
-      $container->get('link_generator'),
-      $container->get('info_parser')
-    );
-  }
 
   /**
    * {@inheritdoc}
@@ -103,10 +38,12 @@ final class AcquiaConnectorForm extends ConfigFormBase implements AcquiaDashboar
     $form['#tree'] = FALSE;
     $module = $this->module;
     $site_name = $this->state->get('spi.site_name');
+    $configured = $this->getProgressState();
     if (!empty($site_name)) {
-      $this->state->set('acquia_connector_progress', TRUE);
+      $configured = TRUE;
+      $this->setState();
     }
-    if ($this->state->get('acquia_connector_progress')) {
+    if ($configured) {
       $form['check_icon'] = [
         '#prefix' => '<span class= "dashboard-check-icon">',
         '#suffix' => "</span>",
@@ -176,31 +113,15 @@ final class AcquiaConnectorForm extends ConfigFormBase implements AcquiaDashboar
     $this->state->set('spi.site_name', $acquia_connector_site_name);
     $this->state->set('acquia_connector_progress', TRUE);
     $this->messenger()->addStatus('The configuration options have been saved.');
+    // Set configuration state for dashboard.
+    $this->setState();
   }
 
   /**
    * {@inheritdoc}
    */
   public function ignoreConfig(array &$form, FormStateInterface $form_state) {
-    $this->state->set('acquia_connector_progress', TRUE);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getModuleStatus() {
-    if ($this->module_handler->moduleExists($this->module)) {
-      return TRUE;
-    }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getProgressState() {
-    if ($this->module_handler->moduleExists($this->module)) {
-      return $this->state->get('acquia_connector_progress');
-    }
+    $this->setState();
   }
 
 }
